# src/Player.Vm.Api.Client/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Player VM API",
    "version": "v1"
  },
  "paths": {
...
    "/api/vms/{id}": {
      "get": {
        "tags": [
          "Vms"
        ],
        "summary": "Retrieve a single Vm by Id",
        "operationId": "getVm",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The Id of the Vm",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Vm"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vm"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vm"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Vm"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vm"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vm"
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Vms"
        ],
        "summary": "Updates a Virtual Machine",
        "description": "Updates a Virtual Machine with the attributes specified\n<para />\nAccessible to a User with management permissions on a team the Virtual Machine is in",
        "operationId": "updateVm",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "The data to update the Virtual Machine with",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VmUpdateForm"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/VmUpdateForm"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/VmUpdateForm"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Vm"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vm"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vm"
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Vms"
        ],
        "summary": "Deletes a Vm",
        "description": "Deletes a Vm with the specified id\n<para />\nAccessible to a User with management permission on a team the Virtual Machine is in",
        "operationId": "deleteVm",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the Vm",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/vms/vsphere/{id}": {
      "get": {
        "tags": [
          "Vsphere"
        ],
        "summary": "Retrieve a single vsphere virtual machine by Id, including a ticket to access it's console",
        "operationId": "getVsphereVirtualMachine",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VsphereVirtualMachine"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VsphereVirtualMachine"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VsphereVirtualMachine"
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/vms/vsphere/{id}/actions/power-on": {
      "post": {
        "tags": [
          "Vsphere"
        ],
        "summary": "Power on a vsphere virtual machine",
        "operationId": "powerOnVsphereVirtualMachine",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/vms/vsphere/{id}/actions/power-off": {
      "post": {
        "tags": [
          "Vsphere"
        ],
        "summary": "Power off a vsphere virtual machine",
        "operationId": "powerOffVsphereVirtualMachine",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/vms/vsphere/{id}/actions/reboot": {
      "post": {
        "tags": [
          "Vsphere"
        ],
        "summary": "Reboot a vsphere virtual machine",
        "operationId": "rebootVsphereVirtualMachine",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/vms/vsphere/{id}/actions/shutdown": {
      "post": {
        "tags": [
          "Vsphere"
        ],
        "summary": "Shutdown a vsphere virtual machine",
        "operationId": "shutdownVsphereVirtualMachine",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/vms/vsphere/{id}/actions/revert": {
      "post": {
        "tags": [
          "Vsphere"
        ],
        "summary": "Revert to the current snapshot of a vsphere virtual machine",
        "operationId": "revertVsphereVirtualMachine",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/vms/vsphere/{id}/snapshots": {
      "get": {
        "tags": [
          "Vsphere"
        ],
        "summary": "Get all snapshots for a vsphere virtual machine",
        "operationId": "getVsphereVirtualMachineSnapshots",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VmSnapshot"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VmSnapshot"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VmSnapshot"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/vms/vsphere/{id}/actions/revert-to-snapshot": {
      "post": {
        "tags": [
          "Vsphere"
        ],
        "summary": "Revert to a specific snapshot of a vsphere virtual machine",
        "operationId": "revertToVsphereVirtualMachineSnapshot",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevertToVsphereVirtualMachineSnapshot"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RevertToVsphereVirtualMachineSnapshot"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RevertToVsphereVirtualMachineSnapshot"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/vms/vsphere/{id}/tools": {
      "get": {
        "tags": [
          "Vsphere"
        ],
        "summary": "Get tools status of a vsphere virtual machine",
        "operationId": "getVsphereVirtualMachineToolsStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VirtualMachineToolsStatus"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VirtualMachineToolsStatus"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VirtualMachineToolsStatus"
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/vms/vsphere/{id}/actions/change-network": {
      "post": {
        "tags": [
          "Vsphere"
        ],
        "summary": "Change the network of a vsphere virtual machine's network adapter",
        "operationId": "changeVsphereVirtualMachineNetwork",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeVsphereVirtualMachineNetwork"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeVsphereVirtualMachineNetwork"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeVsphereVirtualMachineNetwork"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VsphereVirtualMachine"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VsphereVirtualMachine"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VsphereVirtualMachine"
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/vms/vsphere/{id}/actions/validate-credentials": {
      "post": {
        "tags": [
          "Vsphere"
        ],
        "summary": "Validate credentials for a vsphere virtual machine",
        "operationId": "validateVsphereVirtualMachineCredentials",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateVsphereVirtualMachineCredentials"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateVsphereVirtualMachineCredentials"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateVsphereVirtualMachineCredentials"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "RevertToVsphereVirtualMachineSnapshot": {
        "type": "object",
        "properties": {
          "snapshotId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "VmUpdateForm": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "consoleConnectionInfo": {
            "$ref": "#/components/schemas/ConsoleConnectionInfo"
          },
          "proxmoxVmInfo": {
            "$ref": "#/components/schemas/ProxmoxVmInfo"
          },
          "embeddable": {
            "type": "boolean",
            "description": "If false, only allow opening this VM in a new tab"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "http://localhost:5000/connect/authorize",
            "tokenUrl": "http://localhost:5000/connect/token",
            "scopes": {
              "player player-vm": "public api access"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "oauth2": [
        "player player-vm"
      ]
    }
  ],
...
}